@@ -12,23 +12,23 @@
     "devDependencies": {
         "@inertiajs/inertia": "^0.8.4",
         "@inertiajs/inertia-vue3": "^0.3.5",
-        "@inertiajs/progress": "^0.2.4",
+        "@inertiajs/progress": "^0.2.6",
         "@tailwindcss/forms": "^0.2.1",
         "@tailwindcss/typography": "^0.3.0",
         "@vue/compiler-sfc": "3.0.10",
-        "axios": "^0.21",
-        "laravel-mix": "^6.0.19",
+        "axios": "^0.21.4",
+        "laravel-mix": "^6.0.29",
         "lodash": "^4.17.19",
-        "postcss": "^8.1.14",
+        "postcss": "^8.3.6",
         "postcss-import": "^12.0.1",
-        "tailwindcss": "^2.0.1",
-        "vue": "3.0.10",
-        "vue-loader": "^16.1.2"
+        "tailwindcss": "^2.2.14",
+        "vue": "^3.2.11",
+        "vue-loader": "^16.5.0"
     },
     "dependencies": {
-        "@headlessui/vue": "^1.3.0",
-        "@heroicons/vue": "^1.0.2",
-        "@popperjs/core": "^2.9.2",
+        "@headlessui/vue": "^1.4.1",
+        "@heroicons/vue": "^1.0.4",
+        "@popperjs/core": "^2.10.1",
         "vue-multiselect": "^3.0.0-alpha.2"
     }
 }
