@@ -3,8 +3,8 @@
         <div class="md:flex md:flex-col">
             <div class="md:h-screen md:flex md:flex-col">
                 <div class="md:flex md:flex-shrink-0">
-                    <div class="md:flex-shrink-0 md:w-56 px-6 py-4 flex items-center justify-between md:justify-center bg-white dark:bg-gray-700">
-                        <inertia-link class="text-xl text-blue-500 font-bold" href="/">
+                    <div class="md:flex-shrink-0 md:w-56 px-6 py-4 flex items-center justify-between md:justify-center bg-white dark:bg-gray-700 border-r">
+                        <inertia-link class="text-xl text-blue-500 font-bold" :href="route('admin:dashboard')">
                             Admin panel
                         </inertia-link>
                         <div class="-mr-2 flex items-center sm:hidden">
@@ -69,7 +69,7 @@
                     </div>
                 </div>
                 <div class="md:flex md:flex-grow md:overflow-hidden">
-                    <admin-menu class="hidden bg-white dark:bg-gray-700 md:block flex-shrink-0 w-56 overflow-y-auto border-r border-t dark:border-gray-700" />
+                    <admin-menu class="hidden bg-white dark:bg-gray-700 md:block flex-shrink-0 w-56 overflow-y-auto border-r dark:border-gray-700" />
                     <div class="md:flex-1 sm:px-4 py-8 p-1 sm:p-2 md:p-4 md:overflow-y-auto bg-gray-100 dark:bg-gray-800" scroll-region><!-- Page Heading -->
                         <div class="max-w-7xl mx-auto py-5 sm:px-6 lg:px-8">
                             <flash-messages />
